Send reasoning_content back to llamacpp models
The llama.cpp server allows `reasoning_content` to be passed in the
messages history. Normally this will be ignored by most chat templates,
but GPT-OSS and Minimax-M2 were actually trained on conversations
including thinking traces, and perform poorly if it is not present.

This fixes llamacpp models to always send `reasoning_content`, which
will increase the number of LLMs that can be used effectively with
mistral-vibe.